Add basic continuation implemenation to TLS Scheme.
This is the most basic version of our implementation. It gives TLS
scheme access to continuations by guiding them toward scheme's
continuations. Not very satisfying.

The example function doesn't yet work with it, either, because we don't
have ways to handle a sequence of instructions.

Next Step:
Implement the "begin" special form, which will have the form:

  (begin e1 e2 e3 ... en)

It will evaluate each expression in order from left to right, and the
meaning of the expression will be the meaning of "en".

Changes to be committed:
      modified:   tlsschemewithlet.scm