Add more test functions for TLS Scheme.
Having some trouble figuring out basic continatuions. It's possible that
recursion is a bit too fancy for us, so I'm testing that out with two
additional recursive functions that are easier to program. If they work
correctly, then there's problems in my additions to TLS scheme.

Changes to be committed:
      modified:   normal-scheme-test.scm