Fix implemenation of continuations.
I managed to get continuations completely working, at the basic level. I
also left some additional comments for extra clues about what to write
in the report, stuff to focus on. The majority are at the beginning, but
there's some extra comments everywhere there's a modification to TLS
scheme to accommodate for continuations.

Changes to be committed:
      modified:   tlsschemewithlet.scm