Add some more functionality to TLS Scheme/make more tests.
The two rely on each other, as the extra functionality was required for
the extra test.

I realized that another thing that a continuation ought to be able to do
is to return to the end of a chain of deferred function calls, and as we
think about how to create continuations, we should probably keep in mind
how we would do that without taking an entire program snapshot, since
that is not what continuations do.

I wrote a function that tests out this behavior in both TLS Scheme and
normal scheme, and it can be run in either language.

Changes to be committed:
      modified:   normal-scheme-test.scm
      modified:   tlsschemewithlet.scm