Restart work on this project.
The report is finished, but we didn't get very far. I'm going to pick up
this project and continue work on implementing continuations in TLS
Scheme, or in an interpreter very much like it, which is built on an
underlying scheme system.

Roadmap:

- Implement a Continuation Passing Interpreter, built on top of Scheme.
- Implement a call-with-current-continuation function, which will take
the current continuation and make that an argument to a function.
- Implement features that usually use continuations in the interpreter,
like try/catch and maybe some minor threading.
- Examine the memory requirements of continuations when done in this
way, then figure out how it can be done in more efficient ways. Try to
implement one of those ways and compare that to continuations in R5RS
scheme. Compare the performance of this to languages like C and see how
they differ, and why. What stops something like Scheme from being as
efficient as C?

Changes to be committed:
      new file:   EOPL-notes.md
      new file:   scheme-with-continuations.scm