Implement most action functions, except for application and let.
Most of the action functions are complete:

- `*const`
- `*cond`
- `*if`

What is left is:

- `*let`
- `*application`

These two I find hard, because an unknown amount of things have to have
their meanings computed ahead-of-time, before the whole expression can
get its meaning. Evaluating each of those things requires an extension
of context, if I am to evaluate those things eagerly. Lazily would be a
different story.

I also have to figure out how often I need to extend the control context.
It seems fairly silly to extend the control context once for each actual
parameter/name in an a list, because it's not like the evaluation of one
name should return to the evaluation of another name.

The best solution that I have is to "accumulate bindings" until there
are no bindings left to collect, then turn those bindings into a rib,
and extend the current environment, then get the meaning of the
let/non-primitive function body in the extended environment with the
control context from the let/application, since the value of the
let/application is the value of the body evaluated in the augmented
environment.

Feels messy, though.

Changes to be committed:
      modified:   EOPL-notes.md
      modified:   scheme-with-continuations.scm